+++ jobs/cache_maintenance/src/cache_maintenance/queue_metrics.py
@@ -3,7 +3,8 @@
 
 import logging
 
-from libcommon.queue import JobTotalMetricDocument, Queue, WorkerSizeJobsCountDocument
+from libcommon.new_queue.metrics import JobTotalMetricDocument, WorkerSizeJobsCountDocument
+from libcommon.queue import Queue
 
 
 def collect_queue_metrics() -> None:
