+++ libs/libcommon/src/libcommon/new_queue/utils.py
@@ -1,8 +1,9 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2024 The HuggingFace Authors.
 
-from ..queue import JobDocument, JobTotalMetricDocument, WorkerSizeJobsCountDocument
+from ..queue import JobDocument
 from .lock import Lock
+from .metrics import JobTotalMetricDocument, WorkerSizeJobsCountDocument
 
 
 # only for the tests
